+++ changelog.md
@@ -1,5 +1,12 @@
 # changelog
 
+## 3.4.0
+* `CHG` infer `nil` as redundant return value
+  ```lua
+  local function f() end
+  local x = f() -- `x` is `nil` instead of `unknown`
+  ```
+
 ## 3.3.1
 * `FIX` [#1213](https://github.com/sumneko/lua-language-server/issues/1213)
 * `FIX` [#1215](https://github.com/sumneko/lua-language-server/issues/1215)